+++ src/Services/index.js
@@ -5,6 +5,7 @@ export * as locationService from "./locationService";
 export * as feedbackService from "./feedbackService";
 export * as gymService from "./gymService";
 export * as heartRateService from "./heartRateService";
+export * as homeExploreService from "./homeExploreService";
 export * as musicService from "./musicService";
 export * as notificationService from "./notificationService";
 export * as ownWorkoutPostService from "./ownWorkoutPostService";